Linked List - 1.1 - Added reverse() function
This commit adds the reverse function with some added
lines in the main() function as well. The implementation
is quite stable and has been tested quite a bit, as long
as there are no bad inputs the list works perfectly.
------------------------------------------------------
The basics of the reverse() function:
>There are 3 pointers : current,prev,next.
>current=head,prev=NULL Initally.
>the while condition checks when we have reached the end
Inside the loop:
>1st line: Set the next pointer one more than the current
>2nd line: set the current's next pointer to prev
>3rd line: set prev as current shifting it forward
>4th line: st current as next shifting it forward
End result: The pointers move forward and keep changing
the next values of each element, thus reversing the list.